Rebase to master (#3)
* Make _glfwPlatformInitJoysticks() return the correct type in null_joystick

This code was introduced in 782e6b6cefc9ac2a2b6779f9e658ab948809253b.

Closes #1745.

* Update changelog

* Null: Fix out parameters not being set

It turns out platform functions are sometimes called directly instead of
going through the public wrapper.

* Formatting

* Fix formatting in transition guide

* Move Wayland build CI to focal

This finally allows us to drop the Wayland workarounds for CI.

Thanks Travis CI!

* Support interprocedural optimization override

This policy applies the CMAKE_INTERPROCEDURAL_OPTIMIZATION override to
the GLFW target.

Fixes #1753.

* Make old debug context hint/attrib name searchable

* Improve docs scaling on mobile and small screens

Closes #1757.

* Fix typo in GLFWwindowmaximizefun documentation

Thanks brianzuvich!

Fixes #1764.

* Migrate to Sass for Doxygen custom stylesheets

Closes #1759.

* Use consistent indentation for Doxygen custom stylesheet

Most of this file is indented with tabs but parts are indented with spaces.

Closes #1770.

* Update CODEOWNERS for move to SCSS

* Specify maximum version of CMake known to work

Related to #1753.

* Make unsupported versions of CMake fail early

* Wayland: Remove extra-cmake-modules dependency

Fixes #1774.

* Wayland: Remove extra trailing whitespace

* Wayland: Remove extra-cmake-modules from CI

* Docs: Remove mentions of extra-cmake-modules

* Wayland: Remove FindWaylandProtocols.cmake and FindXKBCommon.cmake (not used anymore)

* Win32: Update all monitor handles on config change

Closes #1761.

* Update changelog

Related to #1761.

* Win32: Fix value for _WIN32_WINNT_WINBLUE

An incorrect value for _WIN32_WINNT_WINBLUE caused GLFW to think Windows
8 was 8.1 or greater.  This led to a crash when attempting to call
SetProcessDpiAwareness.

Closes #1775.

* Update changelog and add credit

Related to #1775.

* Update test SPIR-V shader code

Closes #1781.

* Add credit

Related to #1779 and #1781.

* Wayland: Assume 96 DPI if physical size is <= 0

* Fix macOS Print Screen

Many years ago this value was changed from F13 to Print Screen with
1ae9ce1e0a00bf6f8b7e719bf1ec6e73f111afcf.  However, this was later
reverted to F13 with 5759d0fdf272351b35f666718a6e2a665db827e8.

This changes it back to Print Screen for consistency with other
platforms.

Closes #1786.

* Update changelog and add credit

Related to #1786.

* Remove ignored Clang compiler flag

Clang ignores this when C99 is enabled, while GCC does not.  Removing
the checks for Clang to hopefully avoid a false sense of security.

* Win32: Enable /W3 on VS for library sources

/W3 is the default for new VS projects and the library builds cleanly
with it on VS 2010-2019 so let's try to keep it that way.

* Win32: Fix clang-cl interpreting -Wall as /Wall

Unlike -Wall, VS /Wall really means all warnings.

Closes #1780.

* Win32: Filter out duplicate size events

This mirrors the filtering done on X11 and Cocoa.  Possibly this should
be done by shared code instead.

Fixes #1610.

* Wayland: Fix CMake binary dir when built as subdir

Regression introduced by 2747e47393cbca2d09db56223e735bd94b21e2eb.

Related to #1782.

* Fix unquoted paths with variable references

Whitespace is a field separator in CMake.

* Wayland: Ensure paths are passed on in quoted form

* Wayland: Scanner tool has input and output args

* Wayland: Fix indentation

* Wayland: Simplify adding of protocol sources

* Add credit

Related to #1774 and #1782.

* Cocoa: Fix ObjC being built as C with CMake 3.19

CMake 3.19 adds -xc when the LANGUAGE file property is C, breaking our
workaround for CMake 3.15 and earlier not understanding the .m suffix.

Fixes #1787.

* Wayland: Fix destroying CSDs in the correct order

On Wayland we implement Client-Side Decorations if the compositors do not
implement SSDs.  In that case, the destructors of the surfaces were called
in the wrong order, leading to a dereference of an already freed object.
We need to first destroy the subsurface before destroying the parent surface.

Related PR on kitty: https://github.com/kovidgoyal/kitty/pull/3066
Related issue on kitty: https://github.com/kovidgoyal/kitty/issues/3051

Closes #1798.

* Add missing changelog issue number

* Use switch statement instead of if-else-chain for cursor shapes

Closes #1739.

* Wayland: Move 96 DPI default to outputHandleMode

Closes #1792.

* Fix minor typo: If if -> If.

Co-authored-by: Luflosi <luflosi@luflosi.de>
Co-authored-by: Camilla Löwy <elmindreda@elmindreda.org>
Co-authored-by: Siavash Eliasi <siavashserver@gmail.com>
Co-authored-by: Emmanuel Gil Peyrot <linkmauve@linkmauve.fr>
Co-authored-by: Joel Winarske <joel.winarske@gmail.com>
Co-authored-by: Nicolas Caramelli <caramelli.devel@gmail.com>
Co-authored-by: Doug Binks <doug@enkisoftware.com>
Co-authored-by: Vladimir Perminov <snark.gm@gmail.com>
Co-authored-by: Andy Williams <andy@andy.xyz>
Co-authored-by: Camilla Löwy <elmindreda@glfw.org>
Co-authored-by: Leonard König <leonard.r.koenig@googlemail.com>
Co-authored-by: Bhee <me@bheesham.com>